requirements.txt
# Below are for OpenAI requirements

# fastapi
# uvicorn
# langchain
# langchain-community
# openai
# faiss-cpu
# pypdf
# markdown
# beautifulsoup4
# requests

langchain>=0.2.0
langchain-community
langchain-huggingface

# API
fastapi
uvicorn

# RAG pipeline core
transformers
torch
sentence-transformers
faiss-cpu

# Optional (used for preprocessing)
beautifulsoup4
markdown
pypdf
requests
python-dotenv